moved mdxcomponents out and finished ibiomed project posts
