+++ content/projects/dp1.mdx
@@ -1,0 +1,55 @@
+---
+title: "Design Project 1 - GI Joe"
+type: Project
+description: "Designing an improved ostomy bag for a patient."
+date: 2022-10-30
+tags:
+  - IBioMed
+  - Design Project
+  - Year 1
+---
+
+This was the first design project of the year. For this project, we were tasked with improving an ostomy bag based on the patient profile handed to us. Although the outline of the project was given to us, we were given a lot of freedom in how we approached the problem.
+
+## Project Overview
+
+- **Start date:** 2022-09-22
+- **End date:** 2022-10-24
+- **My role:** Subject Matter Expert
+- **Team members:** Aditi Srinivas, Aleen Al Barbarawi, Josiah Kim
+- **Technical skills:** Material sciences, designing, prototyping
+- **Soft skills:** Teamwork, communication, time management, problem solving
+
+## Patient Profile
+
+Our team's patient was Carla Tortelli, an ileostomy patient (the portion of intestine past the ileum was removed/bypassed). She had problems with her one-piece ostomy bag system leaking because of her active lifestyle and how high up the ostomy was placed in her digestive tract. We narrowed down the main issues to be:
+
+- Her output is liquidy, meaning the bag fills up quickly.
+- The sweat from her exercise causes the bag to lose adhesion.
+- The bag can be pretty loud and isn't the most discrete.
+
+## Our Solution
+
+We decided to stick with her one-piece system but we added the following quality of life (QoL) improvements to the bag:
+
+- Increasing the size of the skin patch to increase adhesion.
+- Replacing the skin barrier with a hydrogel adhesive to reduce skin irritation and improve sealing around the stoma.
+- Lining the inside of the bag with ePTFE to reduce noise, increase durability, and reducing leakage.
+
+## My Contributions
+
+As the Subject Matter Expert, I was responsible for compiling the research gathered by our team and formatting it into citations for our report. For technical contributions, I helped with the design of the larger skin patch, as well as suggesting the use of ePTFE to line the inside of the bag. Furthermore, I helped to manufacture the prototype of the bag and produce the presentation and final report.
+
+## Takeaways
+
+One of the biggest takeaways from this project was the importance of communication and learning how to work with people who have different working styles. Some people will naturally be more vocal and take on more of a leadership role, while others will be more reserved and prefer to work in the background. It's important to be able to work with both types of people and to be able to communicate effectively with them.
+
+## Project Reflection
+
+Honestly, I thought that things would have been rougher but it really wasn't too bad. Of course, there were going to be hurdles in communication since everyone in the program has just barely gotten to know each other, but we worked pretty smoothly considering everything that was going on in our busy lives. Though I wouldn't say this was the best design project of the year, I was pretty happy with the result and I'm glad that we were able to get along as a team.
+
+## Featured
+
+![First Slide](/projects/dp-1/title-slide.png)
+
+![LoFi Model](/projects/dp-1/lofi-model.png)
