+++ content/projects/dp2.mdx
@@ -1,0 +1,59 @@
+---
+title: "Design Project 2 - Hips Don't Lie"
+type: Project
+description: "Designing an improved hip implant for a patient."
+date: 2022-12-08
+tags:
+  - IBioMed
+  - Design Project
+  - Year 1
+---
+
+This was the second design project of the year. For this project, we were tasked with designing an improved hip plant based on the patient profile handed to us. The scope and flexibility was similar to the first project, although with less guidance than before.
+
+## Project Overview
+
+- **Start date:** 2022-10-27
+- **End date:** 2022-12-07
+- **My role:** Manager
+- **Team members:** David Segal-Pillemer, Olivia Dmitrovich, Ronav Roy Chowdhury, Safana Al-Emara
+- **Technical skills:** Material sciences, designing, prototyping, CAD in Inventor, Python
+- **Soft skills:** Teamwork, leadership, problem solving, communication
+
+## Patient Profile
+
+Our team's patient was Jackie Chiles, a patient with chronic hip pain since his childhood. Because of the pain, he developed a more sedentary lifestyle, leading to weight gain which worsened the pain. Using patient history and x-ray imagery, we determined that his hip pain was caused by osteoarthritis due to undiagnosed Legg-Calve-Perthes disease. This caused the femoral head to be flattened and the acetabulum to be shallow, leading to a lack of stability in the hip joint. With this, we narrowed down the main issues to be:
+
+- Lack of stability in the hip joint due to the flattened femoral head and shallow acetabulum.
+- Lack of mobility due to the pain and stiffness in the hip joint.
+- Lack of strength due to the sedentary lifestyle.
+
+## Our Solution
+
+We decided to create a custom hip implant named the CHILES (it was a funny backronym). It's a hip implant with the following features:
+
+- A custom-shaped acetabular cup to fit the patient's shallow acetabulum.
+- A custom-designed femoral implant to give the same freedom of movement as his other hip.
+- The femoral implant uses a mixture of porous titanium alloy for bone ingrowth and carbon fibre for strength.
+
+## My Contributions
+
+I was the manager for this project, so I was responsible for the overall management of the project. I created the agenda for each weekly meeting, created our initial Gantt chart, and made sure that everyone was on track with their tasks. Furthermore, I helped to create the poster board for our final presentation and the final report.
+
+For technical contributions, I helped to design the acetabular cup and the plastic spacer. I created the designed in Autodesk Inventor, using the 3D scan of the patient's pelvis as a reference. I also helped to design parts of the femoral implant and the final assembly.
+
+## Takeaways
+
+The biggest takeaway from this project was the importance of staying informed with each subteam and making sure that everyone was on track. I found that during this design project, I was way too focused on my own tasks that I didn't check in on the programming subteam as much as I should've. This reflected in our final prototype, since the programming subteam had difficulties calculating the forces on the hip implant. I feel that had I checked in on them more often, I would have been able to help them with the programming had I known that they were struggling.
+
+## Project Reflection
+
+This was a step up from the first design project in terms of difficulty but I was happy with the results. I liked the increased freedom compared to the first project, as well as the increased challenge of the problem we were trying to solve. I also liked the increased focus on the technical aspects of the project, since I was able to learn more about the design process and the technical aspects of the project. Through this one, I also learned that I have a lot to learn about managing a team, and that I need to be more proactive in checking in on my team members. Overall, I enjoyed this project and I'm looking forward to the next one.
+
+## Featured
+
+![CAD IAM](/projects/dp-2/cad-iam.png)
+
+![3D Model](/projects/dp-2/print.jpg)
+
+![Posterboard](/projects/dp-2/posterboard.png)
