+++ content/projects/dp3.mdx
@@ -1,0 +1,68 @@
+---
+title: "Design Project 3 - A Sense of Independence"
+type: Project
+description: "Designing an accessibility device for power wheelchair users."
+date: 2023-03-08
+tags:
+  - IBioMed
+  - Design Project
+  - Year 1
+---
+
+This was the third design project of the year. This project was the most open-ended of all the projects; we had to find our own demographic, a problem that affects them, and design our own solution. The only stipulation was that we needed to use a sensor, we needed a mechanical component moved with actuators, and we used a Raspberry Pi 4 to control the system.
+
+## Project Overview
+
+- **Start date:** 2023-01-26
+- **End date:** 2023-03-03
+- **My role:** Coordinator
+- **Team members:** Emily Attai, Ethan Otteson, Salma Baig
+- **Technical skills:** Python, unit testing, OOP, Raspberry Pi, CAD in Inventor, 3D printing, laser cutting
+- **Soft skills:** Teamwork, leadership, problem solving, communication, market research, user research, ideation, prototyping
+
+## Demographic and Problem
+
+Given the name of the project, we wanted to focus on a demographic where we could maximize their freedom and independence. Thus, we decided to focus on power wheelchair users affected by neurodegenerative diseases, paralysis, or amputations. We found that many of these users have difficulty controlling their wheelchairs and usually have to rely on caretakers to move them around. Furthermore, solutions like sip-and-puff controllers are extremely expensive, especially for some people who still have some fine motor control.
+
+## Our Solution
+
+We decided to create the BeeAble, a joystick accessory which allows the joystick to be controlled via head gestures. It consists of two parts:
+
+1. The main unit, which goes over the joystick and contains the joystick actuators and the microcontroller.
+2. The head unit, which is worn on the user's head and contains the sensors and haptic feedback circuitry.
+
+Basically, the user makes a series of head gestures which is registered by the orientation sensor on the head unit. The microprocessor then decides which command to execute and sends control signals to the joystick actuator and haptic feedback circuitry.
+
+## My Contributions
+
+I was the coordinator for this project, so I was responsible for taking our meeting minutes during weekly meetings, ensuring clear communication between members, and arranging meetings outside of our design studio as we saw fit.
+
+For technical contributions, I designed, wrote, and tested a large portion of the code as well as the circuitry for the whole system. I designed the following parts of the code:
+
+- Stepper motor driver code.
+- Filtered orientation sensor code.
+- Haptic feedback.
+- Joystick control.
+- Gesture recognition and settings.
+
+## Takeaways
+
+In all honesty, the programming aspect of this design project went about as smooth as it could be. Unit testing all parts of the code was made easy by thoughtfully breaking down the code into small, manageable classes beforehand. The only problem that we had with this project was getting in a rush to finish it all towards the end which is a natural consequence of the first year of university being a bit of a time crunch. Nevertheless, it is still good to be aware of this and try to avoid it as much as possible in the future.
+
+## Project Reflection
+
+Of every design project I have done so far, this one was by far the most enjoyable. I think this was because of the open-ended nature of the project. It was nice to be able to come up with our own problem and solution and not have to worry about meeting a specific set of requirements. Another thing I loved about this project was how much my teammates got along with each other, very quickly. I think this was because we were all very passionate about the project and were all very motivated to do well. I think this project was a great learning experience for me and I am very proud of what we were able to accomplish.
+
+## Links
+
+- [GitHub](https://github.com/IonicArgon/ibehs-dp3-program)
+
+## Featured
+
+<Youtube videoID="oMGegGtGb6E" />
+
+![The Doohickey](/projects/dp-3/doohickey.png)
+
+![The Thingamabob](/projects/dp-3/thingamabob.png)
+
+![The Whatchamacallit](/projects/dp-3/whatchamacallit.png)
