+++ content/projects/vex-2021.mdx
@@ -20,12 +20,4 @@ This is the code I helped write for my team [1104A](https://www.robotevents.com/
 
 ![Robot](/projects/vex-2021/robot.png)
 
-<iframe
-  width="560"
-  height="315"
-  src="https://www.youtube-nocookie.com/embed/OS1HnUw7RkU?si=9QdxwBD00S3KzZuH"
-  title="YouTube video player"
-  frameborder="0"
-  allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share"
-  allowfullscreen
-></iframe>
+<Youtube videoID="OS1HnUw7RkU" />
